+++ app_links/example/ios/Runner/AppDelegate.swift
@@ -2,7 +2,7 @@ import UIKit
 import Flutter
 import app_links
 
-@UIApplicationMain
+@main
 @objc class AppDelegate: FlutterAppDelegate {
   override func application(
     _ application: UIApplication,